pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "course-engine"
version = "0.1.0"
description = "A Python-first, Quarto-backed course production engine aligned to CloudPedagogy AI Capability Framework and Capability-Driven Development."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "CloudPedagogy"}]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "jinja2>=3.1.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
]

[project.scripts]
course-engine = "course_engine.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
